refactoring panic_handler_test files
